allow partial days for expiration days

+++ app/helpers/s3_cleanup_job.rb
@@ -6,7 +6,7 @@ class S3CleanupJob
   end
 
   def perform
-    S3Repository.cleanup(ENV['POET_S3_EXPIRE_DAYS'].to_i)
+    S3Repository.cleanup(ENV['POET_S3_EXPIRE_DAYS'].to_f)
   end
 
 end
